修了修bug

+++ WangPluginPkm/PluginUtil/DisBase/PerfectDitto.cs
@@ -35,123 +35,123 @@ namespace WangPluginPkm.PluginUtil.DisBase
             switch (n)
             {
                 case 0:
-                    pk.Nature = (int)Nature.Bold;
+                    pk.Nature = Nature.Bold;
                     pk.IVs = Atkiv;
                     break;
                 case 1:
-                    pk.Nature = (int)Nature.Timid;
+                    pk.Nature = Nature.Timid;
                     pk.IVs = Spaiv;
                     break;
                 case 2:
-                    pk.Nature = (int)Nature.Adamant;
+                    pk.Nature = Nature.Adamant;
                     pk.IVs = Atkiv;
                     break;
                 case 3:
-                    pk.Nature = (int)Nature.Hasty;
+                    pk.Nature = Nature.Hasty;
                     pk.IVs = Atkiv;
                     break;
                 case 4:
-                    pk.Nature = (int)Nature.Lax;
+                    pk.Nature = Nature.Lax;
                     pk.IVs = Atkiv;
                     break;
                 case 5:
-                    pk.Nature = (int)Nature.Quiet;
+                    pk.Nature = Nature.Quiet;
                     pk.IVs = Tankiv;
                     break;
                 case 6:
-                    pk.Nature = (int)Nature.Rash;
+                    pk.Nature = Nature.Rash;
                     pk.IVs = Atkiv;
                     break;
                 case 7:
-                    pk.Nature = (int)Nature.Mild;
+                    pk.Nature = Nature.Mild;
                     pk.IVs = Atkiv;
                     break;
                 case 8:
-                    pk.Nature = (int)Nature.Modest;
+                    pk.Nature = Nature.Modest;
                     pk.IVs = Spaiv;
                     break;
                 case 9:
-                    pk.Nature = (int)Nature.Lonely;
+                    pk.Nature = Nature.Lonely;
                     pk.IVs = Atkiv;
                     break;
                 case 10:
-                    pk.Nature = (int)Nature.Careful;
+                    pk.Nature = Nature.Careful;
                     pk.IVs = Atkiv;
                     break;
                 case 11:
-                    pk.Nature = (int)Nature.Jolly;
+                    pk.Nature = Nature.Jolly;
                     pk.IVs = Atkiv;
                     break;
                 case 12:
-                    pk.Nature = (int)Nature.Impish;
+                    pk.Nature = Nature.Impish;
                     pk.IVs = Atkiv;
                     break;
                 case 13:
-                    pk.Nature = (int)Nature.Naive;
+                    pk.Nature = Nature.Naive;
                     pk.IVs = Atkiv;
                     break;
                 case 14:
-                    pk.Nature = (int)Nature.Naughty;
+                    pk.Nature = Nature.Naughty;
                     pk.IVs = Atkiv;
                     break;
                 case 15:
-                    pk.Nature = (int)Nature.Calm;
+                    pk.Nature = Nature.Calm;
                     pk.IVs = Atkiv;
                     break;
                 case 16:
-                    pk.Nature = (int)Nature.Gentle;
+                    pk.Nature = Nature.Gentle;
                     pk.IVs = Atkiv;
                     break;
                 case 17:
-                    pk.Nature = (int)Nature.Brave;
+                    pk.Nature = Nature.Brave;
                     pk.IVs = Tiv;
                     break;
                 case 18:
-                    pk.Nature = (int)Nature.Relaxed;
+                    pk.Nature = Nature.Relaxed;
                     pk.IVs = Atkiv;
                     break;
                 case 19:
-                    pk.Nature = (int)Nature.Sassy;
+                    pk.Nature = Nature.Sassy;
                     pk.IVs = Atkiv;
                     break;
                 case 20:
-                    pk.Nature = (int)Nature.Relaxed;
+                    pk.Nature = Nature.Relaxed;
                     pk.IVs = Tankiv;
                     break;
                 case 21:
-                    pk.Nature = (int)Nature.Sassy;
+                    pk.Nature = Nature.Sassy;
                     pk.IVs = Tankiv;
                     break;
                 case 22:
-                    pk.Nature = (int)Nature.Calm;
+                    pk.Nature = Nature.Calm;
                     pk.IVs = Spaiv;
                     break;
                 case 23:
-                    pk.Nature = (int)Nature.Sassy;
+                    pk.Nature = Nature.Sassy;
                     pk.IVs = Tiv;
                     break;
                 case 24:
-                    pk.Nature = (int)Nature.Relaxed;
+                    pk.Nature = Nature.Relaxed;
                     pk.IVs = Tiv;
                     break;
                 case 25:
-                    pk.Nature = (int)Nature.Brave;
+                    pk.Nature = Nature.Brave;
                     pk.IVs = Atkiv;
                     break;
                 case 26:
-                    pk.Nature = (int)Nature.Bold;
+                    pk.Nature = Nature.Bold;
                     pk.IVs = Spaiv;
                     break;
                 case 27:
-                    pk.Nature = (int)Nature.Quiet;
+                    pk.Nature = Nature.Quiet;
                     pk.IVs = Tiv;
                     break;
                 case 28:
-                    pk.Nature = (int)Nature.Quiet;
+                    pk.Nature = Nature.Quiet;
                     pk.IVs = Atkiv;
                     break;
                 case 29:
-                    pk.Nature = (int)Nature.Timid;
+                    pk.Nature = Nature.Timid;
                     pk.IVs = Atkiv;
                     break;
